requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Medical Imaging
monai>=1.3.0
nibabel>=5.0.0
SimpleITK>=2.2.1
pydicom>=2.3.1

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
ipywidgets>=8.0.0

# Data Handling
pandas>=2.0.0
pyyaml>=6.0
h5py>=3.8.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
coloredlogs>=15.0
python-dotenv>=1.0.0

# Uncertainty Quantification
uncertainty-toolbox>=0.1.0

# Image Processing
opencv-python>=4.7.0
albumentations>=1.3.0

# Statistical Analysis
statsmodels>=0.14.0
pingouin>=0.5.3

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Development
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
jupyterlab>=4.0.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
